book page format
made the book only have 1 sheet (which is 2 pages) for now

to add more pages, we can simply copy paste the sheet, and place the sheets in descending order in the hierarchy (the last page should be first in the hierarchy, whereas sheet 1 should be at the bottom). then, the final step is to add each sheet into the "Pages" list in the Book component of the Canvas.

we should also change the aspect ratio, either of the text sprite, or of the sheet, to perfectly match each other